修改错误处理函数使其更易理解，加入SPI_FLASH函数
加入SPI_FLASH的测试代码文件

+++ STM32调试工具/STM32F407_ProjectTemplate/bsp/Response/response.h
@@ -13,7 +13,7 @@ typedef struct
 
 extern usart_data_typed test_data;
 
-int do_process(usart_data_typed* udata);
+uint8_t do_process(usart_data_typed* udata);
 static void do_uart(uint8_t *data);
 static void do_i2c_8RegisterAddr(uint8_t *data);
 static void do_i2c_16RegisterAddr(uint8_t *data);
